+++ Dockerfile
@@ -19,17 +19,20 @@ COPY requirements.txt .
 # Install Python dependencies (including boto3 for AWS integration)
 RUN pip install --no-cache-dir -r requirements.txt boto3
 
-# Copy the providers directory first (for better layer caching)
-COPY tts/providers/ ./providers/
+# Copy shared modules to root
+COPY wer_calculator.py secretmanager.py ./
 
-# Copy main scripts
-COPY tts/runner.py tts/stt_script.py tts/wer_calculator.py ./
+# Copy the entire tts directory
+COPY tts/ ./tts/
+
+# Copy the entire stt directory  
+COPY stt/ ./stt/
+
+# Copy master script
+COPY run_benchmarks.py ./
 
 # Copy environment file (if present)
 COPY .env* ./
-
-# Copy any additional Python modules
-COPY *.py ./
 
 # Create directories for audio files and outputs
 RUN mkdir -p /app/audio_files /app/outputs /app/test_cases
@@ -37,7 +40,7 @@ RUN mkdir -p /app/audio_files /app/outputs /app/test_cases
 # Set environment variables
 ENV PYTHONPATH=/app
 ENV PYTHONUNBUFFERED=1
-ENV AWS_DEFAULT_REGION=us-east-1
+ENV AWS_DEFAULT_REGION=us-east-2
 
-# Default command (can be overridden via EventBridge)
-CMD ["python", "runner.py"]
+# Default command - run the master script
+CMD ["python", "run_benchmarks.py"]
