+++ tts/Test cases.csv
@@ -1,0 +1,31 @@
+Testcase ID,Transcript
+A1,Your order #ORD-24589 shipped via UPS tracking 1Z999AA1234567890 and will arrive between 2:30 PM - 4:45 PM tomorrow.
+A2,"There's a slight delay with your $347.89 order, but we expect it to ship by Friday afternoon."
+A3,"Order status update: your items are now in ""Quality Check"" and should ship within 24-48 hours."
+A4,"Hi Ms. Garcia, your appointment with Dr. Peterson is scheduled for Tuesday, March 5th at 10:30 AM."
+A5,I can reschedule you to either 11:15 AM on Wednesday or - let me check - 2:45 PM on Thursday.
+A6,Please arrive 10 minutes early and bring your insurance card and a photo ID.
+A7,"Let's fix your Wi-Fi: unplug your router for exactly 30 seconds, then wait for the green light."
+A8,"Your device is downloading update version 12.4.1 - about 250 MB remaining, roughly 8 minutes left."
+A9,"I'm seeing error code E-1047 here, which usually means - actually, let me guide you through the solution."
+A10,"For security, please confirm the last 4 digits of the card ending in 8429 used for your recent purchase."
+A11,I need to verify your account: what's the ZIP code we have on file? Please take your time.
+A12,"We detected 2 unusual login attempts from Chicago, Illinois yesterday around 11:30 PM."
+A13,"Hi David, your premium subscription renews in 5 days at $29.99 - would you like to continue or explore other options?"
+A14,"Based on your purchase history, you might like our flash sale: 30% off electronics through Sunday."
+A15,I noticed you haven't logged in since January 15th - is there anything we can help you with?
+A16,"Your invoice shows a balance of $187.50, including the monthly fee plus tax."
+A17,The $24.99 charge is for the premium support package you activated on January 22nd.
+A18,Your auto-payment of $67.50 failed due to an expired card - would you like to update your payment method?
+A19,Thanks for your interest in our software - companies your size typically see 20-25% efficiency gains.
+A20,"For a 50-person team, I'd recommend our Business plan at $149 per month rather than Basic."
+A21,Great timing! We're offering 20% off annual plans - that saves you $359.88 for the first year.
+A22,"On a scale of 1 to 10, how would you rate your current pain level?"
+A23,Have you taken any medications today? Please include prescription drugs and over-the-counter items.
+A24,"Your symptoms sound manageable, but I'd recommend seeing Dr. Martinez within 48 hours to be safe."
+A25,Your PTO request for December 20th through 24th has been approved - that's 3 business days.
+A26,Password reset complete: your temporary password is NewPass2024! Please change it within 24 hours.
+A27,IT ticket about your printer issue has been assigned to technician Lisa Chen.
+A28,"Your wire transfer of $1,247.50 to account ending in 5691 was processed at 2:15 PM."
+A29,"Hi Mr. O'Connor, I'm calling about your consultation - are you available for 20 minutes between 9 AM and noon?"
+A30,"System alert: your server usage hit 87% at 3:42 AM, but it's back to normal levels now."
