+++ wer_calculator.py
